split code into modules, first pass. Rendering tbd
